Se arregla PUT y DELETE de modelos Asignacion y Vencimiento Puntos

+++ src/main/java/py/com/progweb/prueba/ejb/AsignacionDAO.java
@@ -30,7 +30,7 @@ public class AsignacionDAO {
         return  (Integer) q.setParameter("monto", monto).getSingleResult();
     }
 
-    public void deleteAsignacion(Long id_asignacion) {
+    public void deleteAsignacion(Integer id_asignacion) {
         AsignacionPuntos asignacionPuntos= this.em.find(AsignacionPuntos.class,id_asignacion);
         if (asignacionPuntos!=null) {
             this.em.remove(asignacionPuntos);
